Update making-soap-requests.md
Added the SOAPAction header to request header instructions.  My SOAP service would not work without it.  I wrote it succinctly which makes it seem like this is always required, but I couldn't find this written anywhere.  Yours was the best instructions I found, so just added this.  If this is more of a conditional requirement based on the implementation of the SOAP service, I can rephrase to a `may need` instead of `will need`.  But either way, I think the information is valuable.